pyproject.toml
[tool.poetry]
name = "PiCachu"
version = "0.1.0"
description = ""
authors = [
    "... <...>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.10"
requests = "2.27.1"
flask = "2.2"
gunicorn = "20.1.0"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^3.1.0"
flask-cors = "^3.0.10"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.2"
boto3 = "^1.21.42"
uuid6 = "^2022.10.25"
boto3-stubs-lite = "^1.26.8"
opencv-python-headless = "^4.6.0.66"
pillow = "^9.3.0"
scikit-learn = "^1.1.3"
imagehash = "^4.3.1"
dvc = {extras = ["gdrive"], version = "^2.34.2"}
pydrive2 = "^1.15.0"
pyinstaller = "^5.6.2"
transformers = "^4.24.0"
pika = "^1.3.1"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.10.0"
pylint = "^2.15.5"
pylint-quotes = "^0.2.3"
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
pylint-exit = "^1.2.0"

[tool.poe.tasks]
force-cuda11 = "pip3 install torch torchvision --extra-index-url https://download.pytorch.org/whl/cpu"

